+++ tests.py
@@ -2,6 +2,7 @@
 Tests for the sprockets.clients.postgresql package
 
 """
+import datetime
 import mock
 import os
 try:
@@ -9,10 +10,17 @@ try:
 except ImportError:
     import unittest
 
+from tornado import gen
 from sprockets.clients import postgresql
+import queries
+from tornado import testing
 
 
 class TestGetURI(unittest.TestCase):
+
+    def tearDown(self):
+        for key in ['HOST', 'PORT', 'DBNAME', 'USER', 'PASSWORD']:
+            del os.environ['TEST1_%s' % key]
 
     def test_get_uri_returns_proper_values(self):
 
@@ -38,6 +46,10 @@ class TestSession(unittest.TestCase):
         os.environ['TEST2_PASSWORD'] = 'baz'
         self.session = postgresql.Session('test2')
 
+    def tearDown(self):
+        for key in ['HOST', 'PORT', 'DBNAME', 'USER', 'PASSWORD']:
+            del os.environ['TEST2_%s' % key]
+
     def test_session_invokes_queries_session(self):
         self.assertTrue(self.mock_init.called)
 
@@ -54,5 +66,112 @@ class TestTornadoSession(unittest.TestCase):
         os.environ['TEST3_PASSWORD'] = 'baz'
         self.session = postgresql.TornadoSession('test3')
 
+    def tearDown(self):
+        for key in ['HOST', 'PORT', 'DBNAME', 'USER', 'PASSWORD']:
+            del os.environ['TEST3_%s' % key]
+
     def test_session_invokes_queries_session(self):
         self.assertTrue(self.mock_init.called)
+
+
+class SessionIntegrationTests(unittest.TestCase):
+
+    def setUp(self):
+        os.environ['TEST4_HOST'] = 'localhost'
+        os.environ['TEST4_PORT'] = '5432'
+        os.environ['TEST4_DBNAME'] = 'postgres'
+        os.environ['TEST4_USER'] = 'postgres'
+
+        try:
+            self.session = postgresql.Session('test', pool_max_size=10)
+        except postgresql.OperationalError as error:
+            raise unittest.SkipTest(str(error).split('\n')[0])
+
+    def tearDown(self):
+        for key in ['HOST', 'PORT', 'DBNAME', 'USER']:
+            del os.environ['TEST4_%s' % key]
+
+    def test_query_returns_results_object(self):
+        self.assertIsInstance(self.session.query('SELECT 1 AS value'),
+                              queries.Results)
+
+    def test_query_result_value(self):
+        result = self.session.query('SELECT 1 AS value')
+        self.assertDictEqual(result.as_dict(), {'value': 1})
+
+    def test_query_multirow_result_has_at_least_three_rows(self):
+        result = self.session.query('SELECT * FROM pg_stat_database')
+        self.assertGreaterEqual(result.count(), 3)
+
+    def test_callproc_returns_results_object(self):
+        timestamp = int(datetime.datetime.now().strftime('%s'))
+        self.assertIsInstance(self.session.callproc('to_timestamp',
+                                                    [timestamp]),
+                              queries.Results)
+
+    def test_callproc_mod_result_value(self):
+        result = self.session.callproc('mod', [6, 4])
+        self.assertEqual(6 % 4, result[0]['mod'])
+
+
+class TornadoSessionIntegrationTests(testing.AsyncTestCase):
+
+    def setUp(self):
+        super(TornadoSessionIntegrationTests, self).setUp()
+        os.environ['TEST5_HOST'] = 'localhost'
+        os.environ['TEST5_PORT'] = '5432'
+        os.environ['TEST5_DBNAME'] = 'postgres'
+        os.environ['TEST5_USER'] = 'postgres'
+        self.session = postgresql.TornadoSession('test',
+                                                 pool_max_size=10,
+                                                 io_loop=self.io_loop)
+
+    #def tearDown(self):
+    #    for key in ['HOST', 'PORT', 'DBNAME', 'USER']:
+    #        del os.environ['TEST5_%s' % key]
+
+    @testing.gen_test
+    def test_query_returns_results_object(self):
+        try:
+            result = yield self.session.query('SELECT 1 AS value')
+        except postgresql.OperationalError:
+            raise unittest.SkipTest('PostgreSQL is not running')
+        self.assertIsInstance(result, queries.Results)
+        result.free()
+
+    @testing.gen_test
+    def test_query_result_value(self):
+        try:
+            result = yield self.session.query('SELECT 1 AS value')
+        except postgresql.OperationalError:
+            raise unittest.SkipTest('PostgreSQL is not running')
+        self.assertDictEqual(result.as_dict(), {'value': 1})
+        result.free()
+
+    @testing.gen_test
+    def test_query_multirow_result_has_at_least_three_rows(self):
+        try:
+            result = yield self.session.query('SELECT * FROM pg_stat_database')
+        except postgresql.OperationalError:
+            raise unittest.SkipTest('PostgreSQL is not running')
+        self.assertGreaterEqual(result.count(), 3)
+        result.free()
+
+    @testing.gen_test
+    def test_callproc_returns_results_object(self):
+        timestamp = int(datetime.datetime.now().strftime('%s'))
+        try:
+            result = yield self.session.callproc('to_timestamp', [timestamp])
+        except postgresql.OperationalError:
+            raise unittest.SkipTest('PostgreSQL is not running')
+        self.assertIsInstance(result, queries.Results)
+        result.free()
+
+    @testing.gen_test
+    def test_callproc_mod_result_value(self):
+        try:
+            result = yield self.session.callproc('mod', [6, 4])
+        except postgresql.OperationalError:
+            raise unittest.SkipTest('PostgreSQL is not running')
+        self.assertEqual(6 % 4, result[0]['mod'])
+        result.free()