+++ docs/history.rst
@@ -1,4 +1,7 @@
 Version History
 ---------------
+- 1.0.1 [2014-09-03]
+ - Expose psycopg2/queries exceptions, objects, etc from ``sprockets.queries.postgresql``
+ - Add integration testing with PostgreSQL
 - 1.0.0 [2014-08-29]
  - Initial release